+++ R/helper_functions.r
@@ -8,7 +8,6 @@
 #' @param m A matrix.
 
 tidy_matrix <- function(m, symmetric = TRUE) {
-  
   # Check whether there are row and column names
   if (!length(rownames(m)) > 0) {
     stop("Matrix has no row names.")
@@ -40,4 +39,46 @@ tidy_matrix <- function(m, symmetric = TRUE) {
   return(df)   
 }
 
+add_statistic <- function(list, name, value, symbol = NULL, subscript = NULL,
+  interval = NULL, level = NULL, lower = NULL, upper = NULL) {
+  
+  if (!is.null(value)) {
+    if (!is.na(value)) {
+      new_list <- list()
+      new_list$name <- name
+      
+      if (!is.null(symbol)) {
+        if (!is.na(symbol)) new_list$symbol <- symbol  
+      }
+      
+      if (!is.null(subscript)) {
+        if (!is.na(subscript)) new_list$subscript <- subscript  
+      }
+      
+      new_list$value <- value
+      
+      if (!is.null(level)) {
+        if (!is.na(level)) {
+          new_list$interval <- interval
+          new_list$level <- level
+          new_list$lower <- lower
+          new_list$upper <- upper
+        }  
+      }
+      
+      list <- append(list, list(new_list))
+    }
+  }
+  
+  return(list)
+}
+
+add_package_info <- function(list, package) {
+  list$package <- list(
+    name = package,
+    version = getNamespaceVersion(package)[[1]]
+  )
+  
+  return(list)
+}