+++ tests/testthat/test_htest.R
@@ -1,6 +1,9 @@
 
 # Setup -------------------------------------------------------------------
 
+# Load packages
+library(tidystats)
+
 # Load test data
 test_results <- read_stats(system.file("test_data/htest.json", 
   package = "tidystats"))
@@ -16,8 +19,8 @@ test_that("one sample t-tests works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$t_test_one_sample
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -28,8 +31,8 @@ test_that("two sample t-tests works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$t_test_two_sample
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -40,8 +43,8 @@ test_that("Welch t-tests works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$t_test_welch
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -52,8 +55,8 @@ test_that("paired t-tests works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$t_test_paired
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -69,8 +72,8 @@ test_that("pearson correlations works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$correlation_pearson
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -84,8 +87,8 @@ test_that("spearman correlations works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$correlation_spearman
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -99,8 +102,8 @@ test_that("kendall correlations works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$correlation_kendall
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -117,8 +120,8 @@ test_that("pearson's chi-squared tests works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$chi_squared
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -131,8 +134,8 @@ test_that("pearson's chi-squared tests with yates' correction works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$chi_squared_yates
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -145,8 +148,8 @@ test_that("chi-squared tests with for given probabilities works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$chi_squared_prob
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -162,8 +165,8 @@ test_that("wilcoxon signed rank exact tests works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$wilcoxon_signed_rank
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -175,8 +178,8 @@ test_that("wilcoxon rank sum tests with continuity correction works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$wilcoxon_rank_sum_continuity
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -191,8 +194,8 @@ test_that("wilcoxon rank sum tests works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$wilcoxon_rank_sum
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -206,8 +209,8 @@ test_that("wilcoxon rank sum tests works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$wilcoxon_rank_sum_conf
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -222,8 +225,8 @@ test_that("fisher's exact tests works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$fisher_test
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -236,8 +239,8 @@ test_that("fisher's exact tests without a confidence interval works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$fisher_test_no_CI
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -250,8 +253,8 @@ test_that("fisher's exact tests on r x c tables works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$fisher_test_r_by_c
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -267,8 +270,8 @@ test_that("fisher's exact tests with simulated p-value works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$fisher_test_simulated_p
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })
@@ -287,8 +290,8 @@ test_that("fisher's exact tests hybrid works", {
   tidy_model <- tidy_stats(model)
   tidy_model_test <- test_results$fisher_test_hybrid
   
-  tidy_model$package$version <- NULL
-  tidy_model_test$package$version <- NULL
+  tidy_model$package <- NULL
+  tidy_model_test$package <- NULL
   
   expect_equal(tidy_model, tidy_model_test, tolerance = tolerance)
 })